Update __init__.py
Update import_weekly_rosters to check if "age" column was specified by user before computing it for return. NOTE: this will additionally provide a workaround for https://github.com/cooperdff/nfl_data_py/issues/75 for users that do not require "age" in the output, given that column's computation is causing the errors described in that issue.